Implemented DatasetFieldInfoEditForm to simplify editing labels and description at DataSetFieldStructure table.
