+++ odd-platform-ui/src/components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetFieldInfoEditForm/DatasetFieldInfoEditForm.tsx
@@ -1,0 +1,178 @@
+import React from 'react';
+import { Typography } from '@mui/material';
+import { Controller, useFieldArray, useForm } from 'react-hook-form';
+import {
+  DataSetField,
+  DatasetFieldApiUpdateDatasetFieldRequest,
+} from 'generated-sources';
+import DialogWrapper from 'components/shared/DialogWrapper/DialogWrapper';
+import LabelItem from 'components/shared/LabelItem/LabelItem';
+import AppButton from 'components/shared/AppButton/AppButton';
+import AppTextField from 'components/shared/AppTextField/AppTextField';
+import ClearIcon from 'components/shared/Icons/ClearIcon';
+import LabelsAutocompleteContainer from './LabelsAutocomplete/LabelsAutocompleteContainer';
+import * as S from './DatasetFieldInfoEditFormStyles';
+
+interface StructureItemInfoEditFormProps {
+  datasetFieldId: number;
+  datasetFieldFormData: {
+    internalDescription: string;
+    labels: { name: string }[];
+  };
+  isLoading: boolean;
+  updateDataSetFieldFormData: (
+    params: DatasetFieldApiUpdateDatasetFieldRequest
+  ) => Promise<DataSetField>;
+  btnCreateEl: JSX.Element;
+}
+type DatasetFieldInfoFormType = {
+  labels: { name: string }[];
+  internalDescription: string;
+};
+
+const DatasetFieldInfoEditForm: React.FC<StructureItemInfoEditFormProps> = ({
+  datasetFieldId,
+  datasetFieldFormData,
+  isLoading,
+  updateDataSetFieldFormData,
+  btnCreateEl,
+}) => {
+  const methods = useForm<DatasetFieldInfoFormType>({
+    mode: 'all',
+    reValidateMode: 'onChange',
+  });
+
+  const { fields, append, remove } = useFieldArray({
+    control: methods.control,
+    name: 'labels',
+  });
+
+  const initialFormState = { error: '', isSuccessfulSubmit: false };
+  const [{ error, isSuccessfulSubmit }, setFormState] = React.useState<{
+    error: string;
+    isSuccessfulSubmit: boolean;
+  }>(initialFormState);
+
+  const onOpen = (handleOpen: () => void) => () => {
+    methods.reset({
+      labels: datasetFieldFormData.labels?.map(label => ({
+        name: label.name,
+      })),
+    });
+    if (btnCreateEl.props.onClick) btnCreateEl.props.onClick();
+    handleOpen();
+  };
+
+  const clearFormState = () => {
+    setFormState(initialFormState);
+    methods.reset();
+  };
+
+  const handleFormSubmit = (data: DatasetFieldInfoFormType) => {
+    updateDataSetFieldFormData({
+      datasetFieldId,
+      datasetFieldUpdateFormData: {
+        labelNames: data.labels.map(label => label.name),
+        description: data.internalDescription,
+      },
+    }).then(
+      () => {
+        setFormState({ ...initialFormState, isSuccessfulSubmit: true });
+        clearFormState();
+      },
+      (response: Response) => {
+        setFormState({
+          ...initialFormState,
+          error: response.statusText || 'Unable to update info',
+        });
+      }
+    );
+  };
+
+  const handleRemove = (index: number) => () => {
+    remove(index);
+  };
+
+  const formTitle = <Typography variant="h4">Edit information</Typography>;
+
+  const formContent = () => (
+    <form
+      id="dataset-field-info-form"
+      onSubmit={methods.handleSubmit(handleFormSubmit)}
+    >
+      <Typography variant="h5" color="texts.info">
+        Add of edit labels and description
+      </Typography>
+      <Controller
+        name="labels"
+        control={methods.control}
+        render={({ field }) => (
+          <LabelsAutocompleteContainer
+            // eslint-disable-next-line react/jsx-props-no-spreading
+            {...field}
+            appendLabel={append}
+          />
+        )}
+      />
+      <S.LabelItemsContainer sx={{ mt: 1, mb: 1.5 }}>
+        {fields?.map((label, index) => (
+          <LabelItem
+            key={label.id}
+            labelName={label.name}
+            removable
+            unfilled
+            onRemoveClick={handleRemove(index)}
+          />
+        ))}
+      </S.LabelItemsContainer>
+      <Controller
+        control={methods.control}
+        name="internalDescription"
+        defaultValue={datasetFieldFormData.internalDescription || ''}
+        render={({ field }) => (
+          <AppTextField
+            {...field}
+            label="Description"
+            placeholder="Enter description"
+            customEndAdornment={{
+              variant: 'clear',
+              showAdornment: !!field.value,
+              onCLick: () => methods.setValue('internalDescription', ''),
+              icon: <ClearIcon />,
+            }}
+          />
+        )}
+      />
+    </form>
+  );
+
+  const formActionButtons = () => (
+    <>
+      <AppButton
+        size="large"
+        type="submit"
+        form="dataset-field-info-form"
+        color="primary"
+        fullWidth
+      >
+        Save
+      </AppButton>
+    </>
+  );
+
+  return (
+    <DialogWrapper
+      renderOpenBtn={({ handleOpen }) =>
+        React.cloneElement(btnCreateEl, { onClick: onOpen(handleOpen) })
+      }
+      title={formTitle}
+      renderContent={formContent}
+      renderActions={formActionButtons}
+      handleCloseSubmittedForm={isSuccessfulSubmit}
+      isLoading={isLoading}
+      errorText={error}
+    />
+  );
+};
+
+export default DatasetFieldInfoEditForm;
