+++ odd-platform-ui/src/components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetFieldInfoEditForm/DatasetFieldInfoEditFormStyles.ts
@@ -1,0 +1,8 @@
+import { styled } from '@mui/material/styles';
+import { Box } from '@mui/material';
+
+export const LabelItemsContainer = styled(Box)(({ theme }) => ({
+  paddingBottom: theme.spacing(1.5),
+  borderBottom: '1px solid',
+  borderBottomColor: theme.palette.backgrounds.primary,
+}));
