+++ odd-platform-ui/src/components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetFieldInfoEditForm/LabelsAutocomplete/LabelsAutocompleteContainer.tsx
@@ -1,7 +1,7 @@
 import { connect } from 'react-redux';
 import { RootState } from 'redux/interfaces';
 import { fetchLabelsList } from 'redux/thunks/labels.thunks';
-import LabelsAutocomplete from './LabelsAutocomplete';
+import LabelsAutocomplete from 'components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetFieldInfoEditForm/LabelsAutocomplete/LabelsAutocomplete';
 
 const mapStateToProps = (state: RootState) => ({});
