+++ odd-platform-ui/src/components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetStructureItem.tsx
@@ -16,7 +16,6 @@ import { format } from 'date-fns';
 import NumberFormatted from 'components/shared/NumberFormatted/NumberFormatted';
 import LabeledInfoItem from 'components/shared/LabeledInfoItem/LabeledInfoItem';
 import LabelItem from 'components/shared/LabelItem/LabelItem';
-import KebabIcon from 'components/shared/Icons/KebabIcon';
 import PlusIcon from 'components/shared/Icons/PlusIcon';
 import MinusIcon from 'components/shared/Icons/MinusIcon';
 import LineBreakIcon from 'components/shared/Icons/LineBreakIcon';
@@ -24,6 +23,8 @@ import InformationIcon from 'components/shared/Icons/InformationIcon';
 import DatasetStructureFieldTypeLabel from 'components/DataEntityDetails/DatasetStructure/DatasetStructureFieldTypeLabel/DatasetStructureFieldTypeLabel';
 import AppTooltip from 'components/shared/AppTooltip/AppTooltip';
 import AppIconButton from 'components/shared/AppIconButton/AppIconButton';
+import AppButton from 'components/shared/AppButton/AppButton';
+import DatasetFieldInfoEditFormContainer from 'components/DataEntityDetails/DatasetStructure/DatasetStructureTable/DatasetStructureItem/DatasetFieldInfoEditForm/DatasetFieldInfoEditFormContainer';
 import { StylesType } from './DatasetStructureItemStyles';
 
 interface DatasetStructureItemProps extends StylesType {
@@ -201,27 +202,14 @@ const DatasetStructureItem: React.FC<DatasetStructureItemProps> = ({
             </Grid>
             <Grid item className={classes.typeCol}>
               <div className={classes.optionsBtn}>
-                <AppTooltip
-                  control="byClick"
-                  renderContent={() => (
-                    <>
-                      {/* <LabelsEditFormContainer */}
-                      {/*  datasetFieldId={datasetField.id} */}
-                      {/*  btnCreateEl={<MenuItem>Edit Labels</MenuItem>} */}
-                      {/* /> */}
-                      {/* <InternalDescriptionFormDialogContainer */}
-                      {/*  datasetFieldId={datasetField.id} */}
-                      {/*  btnCreateEl={<MenuItem>Edit Description</MenuItem>} */}
-                      {/* /> */}
-                    </>
-                  )}
-                >
-                  <AppIconButton
-                    size="medium"
-                    color="primaryLight"
-                    icon={<KebabIcon />}
-                  />
-                </AppTooltip>
+                <DatasetFieldInfoEditFormContainer
+                  datasetFieldId={datasetField.id}
+                  btnCreateEl={
+                    <AppButton size="medium" color="primaryLight">
+                      Edit
+                    </AppButton>
+                  }
+                />
               </div>
               <DatasetStructureFieldTypeLabel
                 typeName={datasetField.type.type}
