--- odd-platform-ui/src/components/DataEntityDetails/DatasetStructure/InternalDescriptionFormDialog/InternalDescriptionFormDialogStyles.ts
@@ -1,11 +1,0 @@
-import { Theme } from '@mui/material';
-
-import { WithStyles } from '@mui/styles';
-import createStyles from '@mui/styles/createStyles';
-
-export const styles = (theme: Theme) =>
-  createStyles({
-    form: {},
-  });
-
-export type StylesType = WithStyles<typeof styles>;
